separate icon utils
